+++ app/src/main/java/com/ayuget/redface/data/api/MDService.java
@@ -145,5 +145,5 @@ public interface MDService {
     /**
      * Searches a particular word and/or author in a topic, starting at a given post id.
      */
-    Observable<TopicSearchResult> searchInTopic(User user, Topic topic, int startFromPostId, String word, String author, boolean firstSearch);
+    Observable<TopicSearchResult> searchInTopic(User user, Topic topic, long startFromPostId, String word, String author, boolean firstSearch);
 }
